api/pyproject.toml
[project]
name = "pipeline-as-art-api"
version = "0.1.0"
description = "FastAPI service for Pipeline as Art"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pyyaml>=6.0.1"
]

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
workers = 1
